GVT-1731 Reorganize styles
