+++ ui/src/data-products/element-list/element-table.scss
@@ -1,0 +1,23 @@
+.element-table {
+    &__table-container {
+        overflow: auto;
+        padding-left: 32px;
+    }
+
+    &__element-count {
+        padding-left: 32px;
+    }
+
+    &__table-heading {
+        position: sticky;
+        top: 0;
+
+        th {
+            min-width: 100px;
+        }
+    }
+
+    &__column--number {
+        text-align: right;
+    }
+}
