+++ ui/src/data-products/element-list/plan-geometry-search.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { useTranslation } from 'react-i18next';
-import styles from 'data-products/element-list/element-list-view.scss';
+import styles from 'data-products/data-product-view.scss';
 import { FieldLayout } from 'vayla-design-lib/field-layout/field-layout';
 import { Dropdown } from 'vayla-design-lib/dropdown/dropdown';
 import { Checkbox } from 'vayla-design-lib/checkbox/checkbox';
@@ -109,7 +109,7 @@ const PlanGeometrySearch = ({
 
     return (
         <React.Fragment>
-            <p className={styles['element-list__geometry-search-legend']}>
+            <p className={styles['data-product__search-legend']}>
                 {t('data-products.element-list.plan-legend')}
             </p>
             <div className={styles['element-list__geometry-search']}>
@@ -120,7 +120,7 @@ const PlanGeometrySearch = ({
                             {planSources.map((source) => (
                                 <span
                                     key={source.value}
-                                    className={styles['element-list-view__radio-layout']}>
+                                    className={styles['data-product-view__radio-layout']}>
                                     <Radio
                                         checked={state.source === source.value}
                                         onChange={() => setSource(source.value)}>
