+++ ui/src/data-products/vertical-geometry/location-track-vertical-geometry-search.tsx
@@ -1,12 +1,12 @@
 import * as React from 'react';
 import { useTranslation } from 'react-i18next';
-import styles from 'data-products/vertical-geometry/vertical-geometry-view.scss';
+import styles from 'data-products/data-product-view.scss';
 
 export const LocationTrackVerticalGeometrySearch: React.FC = () => {
     const { t } = useTranslation();
     return (
         <React.Fragment>
-            <p className={styles['vertical-geometry__search-legend']}>
+            <p className={styles['data-product__search-legend']}>
                 {t('data-products.vertical-geometry.location-track-search-legend')}
             </p>
         </React.Fragment>
