--- ui/src/data-products/vertical-geometry/vertical-geometry-view.scss
@@ -1,20 +1,0 @@
-.vertical-geometry-view {
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-
-    &__radio-layout > .radio {
-        margin-right: 20px;
-        display: inline-block;
-        margin-top: 8px;
-        margin-bottom: 20px;
-    }
-
-    &__header-container {
-        padding-left: 32px;
-    }
-}
-
-.vertical-geometry__search-legend {
-    max-width: 800px;
-}
